Add mappings for non-vocab resources (libraries) and use them in disambiguation
